+++ components/Conversations/Detail/ConversationDetail/NoAudioDetected.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
-import GreyShieldTick from '../Icon/GreyShieldTick'
-import BigSoundIcon from '../Icon/BigSoundIcon'
+import GreyShieldTick from '../Icon/DetailIcons/NoAudioDetected/GreyShieldTick'
+import BigSoundIcon from '../Icon/DetailIcons/NoAudioDetected/BigSoundIcon'
 import { Trash } from 'iconsax-react'
 
 const NoAudioDetected: React.FC = () => {
