+++ components/Conversations/Detail/Icon/CopyIcon.tsx
@@ -1,10 +1,10 @@
-import React from 'react';
+import React from 'react'
 
 const CopyIcon: React.FC = () => (
     <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" viewBox="0 0 16 16" fill="none">
         <path d="M10.6666 8.59967V11.3997C10.6666 13.733 9.73325 14.6663 7.39992 14.6663H4.59992C2.26659 14.6663 1.33325 13.733 1.33325 11.3997V8.59967C1.33325 6.26634 2.26659 5.33301 4.59992 5.33301H7.39992C9.73325 5.33301 10.6666 6.26634 10.6666 8.59967Z" fill="#484848"/>
         <path d="M11.4001 1.33301H8.60009C6.67939 1.33301 5.71176 1.96994 5.42932 3.49228C5.32773 4.03984 5.79508 4.49967 6.35199 4.49967H7.40009C10.2001 4.49967 11.5001 5.79967 11.5001 8.59967V9.64777C11.5001 10.2047 11.9599 10.672 12.5075 10.5704C14.0298 10.288 14.6668 9.32037 14.6668 7.39967V4.59967C14.6668 2.26634 13.7334 1.33301 11.4001 1.33301Z" fill="#484848"/>
     </svg>
-);
+)
 
-export default CopyIcon; 
+export default CopyIcon
