+++ components/Conversations/Detail/Icon/DetailIcons/ActiveConversationIcon/BigGreenSoundIcon.tsx
