+++ components/Conversations/Detail/Icon/DetailIcons/ConversationDetailIcon/GreyClipboardIcon.tsx
