+++ components/Conversations/Detail/Icon/DetailIcons/NoAudioDetected/GreyShieldTick.tsx
