+++ components/Conversations/Detail/Icon/DetailIcons/TranscriptionDisabled/ShieldTickIcon.tsx
