+++ components/Conversations/Detail/Icon/FlagIcon.tsx
@@ -1,10 +1,10 @@
-import React from 'react';
+import React from 'react'
 
 const FlagIcon: React.FC = () => (
 <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" viewBox="0 0 16 16" fill="none">
     <path d="M12.0134 8.21967L11.2 7.40634C11.0067 7.23967 10.8934 6.99301 10.8867 6.71967C10.8734 6.41967 10.9933 6.11967 11.2133 5.89967L12.0134 5.09967C12.7067 4.40634 12.9667 3.73967 12.7467 3.21301C12.5334 2.69301 11.8734 2.40634 10.9 2.40634H3.93335V1.83301C3.93335 1.55967 3.70668 1.33301 3.43335 1.33301C3.16002 1.33301 2.93335 1.55967 2.93335 1.83301V14.1663C2.93335 14.4397 3.16002 14.6663 3.43335 14.6663C3.70668 14.6663 3.93335 14.4397 3.93335 14.1663V10.913H10.9C11.86 10.913 12.5067 10.6197 12.7267 10.093C12.9467 9.56634 12.6934 8.90634 12.0134 8.21967Z" fill="#484848"/>
 </svg>
-);
+)
 
-export default FlagIcon; 
+export default FlagIcon
