+++ components/Conversations/Detail/Icon/SaveAddIcon.tsx
@@ -1,10 +1,10 @@
-import React from 'react';
+import React from 'react'
 
 const SaveAddIcon: React.FC = () => (
     <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" viewBox="0 0 16 16" fill="none">
         <path d="M14.6666 3.40634V10.9797C14.6666 11.9463 13.9733 12.353 13.1266 11.8863L11.84 11.1663C11.7333 11.1063 11.6666 10.993 11.6666 10.873V5.99301C11.6666 4.29967 10.2866 2.91967 8.59331 2.91967H5.87998C5.63331 2.91967 5.45998 2.65967 5.57331 2.44634C5.91998 1.78634 6.61331 1.33301 7.40664 1.33301H12.5933C13.7333 1.33301 14.6666 2.26634 14.6666 3.40634Z" fill="#484848"/>
         <path d="M8.59325 3.91992H3.40659C2.26659 3.91992 1.33325 4.85325 1.33325 5.99326V13.5666C1.33325 14.5333 2.02659 14.9399 2.87325 14.4733L5.49325 13.0133C5.77325 12.8599 6.22659 12.8599 6.50659 13.0133L9.12659 14.4733C9.97325 14.9399 10.6666 14.5333 10.6666 13.5666V5.99326C10.6666 4.85325 9.73325 3.91992 8.59325 3.91992ZM7.33325 8.49992H6.49992V9.33326C6.49992 9.60659 6.27325 9.83326 5.99992 9.83326C5.72659 9.83326 5.49992 9.60659 5.49992 9.33326V8.49992H4.66659C4.39325 8.49992 4.16659 8.27326 4.16659 7.99992C4.16659 7.72659 4.39325 7.49992 4.66659 7.49992H5.49992V6.66659C5.49992 6.39326 5.72659 6.16659 5.99992 6.16659C6.27325 6.16659 6.49992 6.39326 6.49992 6.66659V7.49992H7.33325C7.60659 7.49992 7.83325 7.72659 7.83325 7.99992C7.83325 8.27326 7.60659 8.49992 7.33325 8.49992Z" fill="#484848"/>
     </svg>
-);
+)
 
-export default SaveAddIcon; 
+export default SaveAddIcon
