+++ components/Conversations/Detail/Icon/SendIcon.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React from 'react'
 
 const SendIcon: React.FC = () => (
   <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" viewBox="0 0 16 16" fill="none">
@@ -6,4 +6,4 @@ const SendIcon: React.FC = () => (
   </svg>
 )
 
-export default SendIcon; 
+export default SendIcon
