+++ components/Conversations/Settings/SettingsPage.tsx
@@ -23,8 +23,6 @@ const SettingsPage: React.FC = () => {
   const { isAudioOn, setIsAudioOn, conversations } = useConversations()
   // Audio Transcription State
   const [audioState, setAudioState] = useState<AudioState>('inactive')
-  const [asrDuration, setAsrDuration] = useState<number>(2)
-  const [isCloudTranscriptOn, setIsCloudTranscriptOn] = useState<boolean>(true)
   const { deleteAllConversations } = useConversations()
 
   // Initialize audioState based on isAudioOn
@@ -32,19 +30,6 @@ const SettingsPage: React.FC = () => {
     setAudioState(isAudioOn ? 'active' : 'off')
   }, [isAudioOn])
 
-  // Options for the dropdown (box 3)
-  const asrDurationOptions = useMemo(() => {
-    return [
-      { label: '2 hours', value: 2 },
-      { label: '4 hours', value: 4 },
-      { label: '8 hours', value: 8 },
-      { label: '12 hours', value: 12 },
-      { label: '24 hours', value: 24 }
-    ]
-  }, [])
-  const handleAsrDurationChange = (option: { value: number }) => {
-    setAsrDuration(option.value)
-  }
 
   // Toggle Audio Transcription
   const handleToggle = () => {
@@ -55,10 +40,6 @@ const SettingsPage: React.FC = () => {
 
   const getSoundIconColor = () => {
     return audioState === 'active' ? '#4CEDA0' : '#484848'
-  }
-
-  const handleCloudToggle = () => {
-    setIsCloudTranscriptOn(!isCloudTranscriptOn)
   }
 
   return (
@@ -136,71 +117,6 @@ const SettingsPage: React.FC = () => {
           </div>
         </>
       )}
-
-      {/* Audio Transcript Duration */}
-      {/* <div className="mb-8 flex flex-col gap-px">
-        <div className="flex items-center justify-between overflow-hidden rounded-t-2xl bg-white/[0.02] px-6 py-3 pr-3">
-          <div className="font-inter text-[15px] font-medium leading-normal text-[#eeeeee]">
-            Audio Transcript Duration
-          </div>
-          <div>
-            <Dropdown
-              key={asrDuration}
-              value={asrDuration}
-              onSelect={handleAsrDurationChange}
-              options={asrDurationOptions}
-              style={{ minWidth: '100px' }}
-            />
-          </div>
-        </div>
-
-        <div className="rounded-b-2xl bg-white/[0.02] px-6 py-4">
-          <div className="font-inter shrink grow basis-0 text-[15px] font-normal leading-normal text-[#b4b4b4] opacity-50">
-            Length of audio transcript duration that will be stored in memory. In the interest of your privacy, there is
-            no data saved anywhere.
-          </div>
-        </div>
-      </div> */}
-
-      {/* Cloud Transcript */}
-
-      {/* <div className="flex flex-col gap-px mb-8">
-        <div className="flex justify-between items-center py-3 px-6 pr-3 bg-white/[0.02] rounded-t-2xl overflow-hidden">
-          <div className="text-[#eeeeee] text-[15px] font-medium font-inter leading-normal">Cloud Transcript</div>
-          <div className="flex items-center gap-2">
-            <button
-              onClick={handleCloudToggle}
-              className="flex items-center gap-1.5 text-xs font-publicsans"
-            >
-              <div
-                className={`text-right ${
-                  isCloudTranscriptOn ? 'text-[#00cc88]' : 'text-white/40'
-                } text-xs font-normal font-['Public Sans'] leading-snug`}
-              >
-                {isCloudTranscriptOn ? 'ON' : 'OFF'}
-              </div>
-              <div className="w-[49px] h-[26px] relative rounded-2xl">
-                <div
-                  className={`w-[49px] h-[26px] left-0 top-0 absolute ${
-                    isCloudTranscriptOn ? 'bg-[#00cc88]' : 'bg-black'
-                  } rounded-full`}
-                />
-                <div
-                  className={`w-6 h-6 absolute ${
-                    isCloudTranscriptOn ? 'left-[24px] bg-white' : 'left-[1px] bg-white/40'
-                  } top-[1px] rounded-full shadow`}
-                />
-              </div>
-            </button>
-          </div>
-        </div>
-
-        <div className="px-6 py-4 bg-white/[0.02] rounded-b-2xl">
-          <div className="grow shrink basis-0 opacity-50 text-[#b4b4b4] text-[15px] font-normal font-inter leading-normal">
-            Allow transcription to work in the cloud whenever your device is unable to transcribe your conversations locally. No audio or text transcription is stored anywhere to protect your privacy.
-          </div>
-        </div>
-      </div> */}
 
       {/* Delete Button */}
       {conversations.length > 0 && (
